Auth done

+++ app/Http/Controllers/NewsController.php
@@ -7,7 +7,6 @@ use App\Contracts\Repositories\NewsRepositoryContract;
 use App\Contracts\Services\NewsServiceContract;
 use App\Http\Requests\News\StoreRequest;
 use App\Http\Requests\News\UpdateRequest;
-use App\Http\Requests\Tag\StoreRequest as TagRequest;
 use App\Http\Resources\NewsResource;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
@@ -33,17 +32,21 @@ class NewsController extends Controller
 
     public function store(StoreRequest $request)
     {
-        return new NewsResource($this->newsService->create($request->validated()));
+        return new NewsResource($this->newsService->create(
+            $request->validated(),
+            $request->user()->id)
+        );
     }
 
     public function update(int $id,UpdateRequest $request)
     {
+        $this->authorize('update', $this->newsRepository->find($id));
         return $this->newsService->update($request->validated(),$id);
-
     }
 
     public function destroy(int $id)
     {
+        $this->authorize('destroy', $this->newsRepository->find($id));
         return $this->newsRepository->destroy($id);
     }
 }
